deprecated_and_misc/deprecated_requirements.txt
# yolov5 basic requirements
# for more info see yolov5/requirements.txt
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
pillow
PyYAML>=5.3
scipy>=1.4.1
tensorboard>=2.2
torch>=1.6.0
torchvision>=0.7.0
tqdm>=4.41.0

# rest of the pipeline 
voila
pandas
#scikit-learn
ipywidgets
gtts
ipykernel
google-cloud-vision
